WebApp/requirements.txt
absl-py==0.13.0
asgiref==3.2.10
astroid==2.4.2
astunparse==1.6.3
autopep8==1.5.3
botocore==1.20.16
Brotli==1.0.9
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
click==7.1.2
cloudinary==1.25.0
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
dash==1.20.0
dash-core-components==1.16.0
dash-html-components==1.1.3
dash-renderer==1.9.1
dash-table==4.11.3
defusedxml==0.6.0
diff-match-patch==20200713
dj3-cloudinary-storage==0.0.4
Django==3.1.4
django-address==0.2.5
django-allauth==0.42.0
django-appconf==1.0.4
django-ckeditor==5.9.0
django-cloudinary-storage==0.3.0
django-cors-headers==3.4.0
django-countries==7.1
django-filter==2.4.0
django-imagekit==4.0.2
django-js-asset==1.2.2
django-plotly-dash==1.6.4
django-rest-auth==0.9.5
django-rest-passwordreset==1.1.0
django-storages==1.11.1
django-webpack-loader==0.6.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
dpd-components==0.1.0
dynaconf==3.0.0
et-xmlfile==1.0.1
Flask==2.0.1
Flask-Compress==1.10.1
flatbuffers==1.12
future==0.18.2
gast==0.4.0
google-api-core==1.23.0
google-api-python-client==1.12.8
google-auth==1.31.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
googleapis-common-protos==1.52.0
grpcio==1.34.1
gunicorn==20.0.4
h5py==3.1.0
httplib2==0.18.1
hvac==0.10.4
idna==2.10
isort==4.3.21
itsdangerous==2.0.1
itypes==1.2.0
jdcal==1.4.1
Jinja2==3.0.1
jmespath==0.10.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
lazy-object-proxy==1.4.3
loguru==0.5.1
Markdown==3.3.4
MarkupPy==1.14
MarkupSafe==2.0.1
mccabe==0.6.1
numpy==1.19.5
oauth2client==4.1.3
oauthlib==3.1.1
odfpy==1.4.1
openapi-codec==1.3.2
opencv-python==4.5.2.54
openpyxl==3.0.5
opt-einsum==3.3.0
pandas==1.0.5
pilkit==2.0
Pillow==7.2.0
plotly==5.1.0
protobuf==3.17.3
psycopg2==2.8.5
py3dns==3.2.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
PyJWT==1.7.1
pylint==2.5.3
pylint-plugin-utils==0.6
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.1
PyYAML==5.3.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
shortuuid==1.0.1
simplejson==3.17.2
six==1.15.0
sqlparse==0.3.1
tablib==2.0.0
tenacity==7.0.0
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
toml==0.10.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.5
validate-email==1.3
Werkzeug==2.0.1
whitenoise==5.2.0
win32-setctime==1.0.1
wrapt==1.12.1
xlrd==1.2.0
xlwt==1.3.0
